Fixed BilinearInterpolationScaling
I was casting non null pixels to bytes by accident

+++ src/filters/Homework1/BilinearInterpolationFilter.cs
@@ -135,7 +135,7 @@ namespace CS555.Homework1
 					}
 					else
 					{
-						outLine[j] = (byte)val; 
+						outLine[j] = (int)val; 
 					}
 				}
 			}
